requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
pillow>=9.0.0
cvzone>=1.5.0
requests>=2.28.0
pyautogui>=0.9.54
flask>=3.0.0
